Moved comm messages creators to messaging package
