+++ packages/messaging/src/index.ts
@@ -24,6 +24,62 @@ export function createExecuteRequest(code: string = ""): ExecuteRequest {
 }
 
 /**
+ * creates a comm open message
+ * @param  {string} comm_id       uuid
+ * @param  {string} target_name   comm handler
+ * @param  {any} data             up to the target handler
+ * @param  {string} target_module [Optional] used to select a module that is responsible for handling the target_name
+ * @return {jmp.Message}          Message ready to send on the shell channel
+ */
+export function createCommOpenMessage(
+  comm_id: string,
+  target_name: string,
+  data: any = {},
+  target_module: string
+) {
+  const msg = createMessage("comm_open", {
+    content: { comm_id, target_name, data }
+  });
+  if (target_module) {
+    msg.content.target_module = target_module;
+  }
+  return msg;
+}
+
+/**
+ * creates a comm message for sending to a kernel
+ * @param  {string}     comm_id    unique identifier for the comm
+ * @param  {Object}     data       any data to send for the comm
+ * @param  {Uint8Array} buffers    arbitrary binary data to send on the comm
+ * @return {jmp.Message}           jupyter message for comm_msg
+ */
+export function createCommMessage(
+  comm_id: string,
+  data: any = {},
+  buffers: Uint8Array = new Uint8Array([])
+) {
+  return createMessage("comm_msg", { content: { comm_id, data }, buffers });
+}
+
+/**
+ * creates a comm close message for sending to a kernel
+ * @param  {Object} parent_header    header from a parent jupyter message
+ * @param  {string}     comm_id      unique identifier for the comm
+ * @param  {Object}     data         any data to send for the comm
+ * @return {jmp.Message}             jupyter message for comm_msg
+ */
+export function createCommCloseMessage(
+  parent_header: any,
+  comm_id: string,
+  data: any = {}
+) {
+  return createMessage("comm_close", {
+    content: { comm_id, data },
+    parent_header
+  });
+}
+
+/**
  * operator for getting all messages that declare their parent header as
  * parentMessage's header.
  *
@@ -71,16 +127,23 @@ export function childOf(
  * @returns A function that takes an Observable of kernel messages and returns
  * messages that have the given comm id
  */
-export function withCommId(comm_id: string): (source: Observable<JupyterMessage<MessageType, any>>) => any  {
+export function withCommId(
+  comm_id: string
+): (source: Observable<JupyterMessage<MessageType, any>>) => any {
   return (source: Observable<JupyterMessage>) => {
-      return Observable.create((subscriber: Subscriber<JupyterMessage>) => source.subscribe(msg => {
+    return Observable.create((subscriber: Subscriber<JupyterMessage>) =>
+      source.subscribe(
+        msg => {
           if (msg && msg.content && msg.content.comm_id === comm_id) {
-              subscriber.next(msg);
+            subscriber.next(msg);
           }
-      }, 
-      // be sure to handle errors and completions as appropriate and
-      // send them along
-      err => subscriber.error(err), () => subscriber.complete()));
+        },
+        // be sure to handle errors and completions as appropriate and
+        // send them along
+        err => subscriber.error(err),
+        () => subscriber.complete()
+      )
+    );
   };
 }
